renamed packages to follow standard package naming conventions
