--- src/main/java/com/example/demo/Actors/Destructible.java
@@ -1,9 +1,0 @@
-package com.example.demo.Actors;
-
-public interface Destructible {
-
-    void takeDamage();
-
-    void destroy();
-
-}
