--- src/main/java/com/example/demo/ImageInitializers/WinImage.java
@@ -1,27 +1,0 @@
-package com.example.demo.ImageInitializers;
-
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-
-import java.util.Objects;
-
-public class WinImage extends ImageView {
-
-    private static final String IMAGE_NAME = "/com/example/demo/images/youwin.png";
-    private static final int HEIGHT = 500;
-    private static final int WIDTH = 600;
-
-    public WinImage(double xPosition, double yPosition) {
-        this.setImage(new Image(Objects.requireNonNull(getClass().getResource(IMAGE_NAME)).toExternalForm()));
-        this.setVisible(false);
-        this.setFitHeight(HEIGHT);
-        this.setFitWidth(WIDTH);
-        this.setLayoutX(xPosition);
-        this.setLayoutY(yPosition);
-    }
-
-    public void showWinImage() {
-        this.setVisible(true);
-    }
-
-}
