--- src/main/java/com/example/demo/Levels/LevelFive.java
@@ -1,47 +1,0 @@
-package com.example.demo.Levels;
-
-import com.example.demo.Planes.Boss;
-import com.example.demo.LevelViews.LevelView;
-import com.example.demo.LevelViews.LevelViewBoss;
-
-public class LevelFive extends LevelParent {
-
-    private static final String BACKGROUND_IMAGE_NAME = "/com/example/demo/images/background2.jpg";
-    private static final String NEXT_LEVEL = "com.example.demo.Levels.LevelThree";
-    private static final int PLAYER_INITIAL_HEALTH = 5;
-    private final Boss boss;
-    private LevelViewBoss levelView;
-
-    public LevelFive(double screenHeight, double screenWidth) {
-        super(BACKGROUND_IMAGE_NAME, screenHeight, screenWidth, PLAYER_INITIAL_HEALTH);
-        boss = new Boss(levelView);
-    }
-
-    @Override
-    protected void initializeFriendlyUnits() {
-        getRoot().getChildren().add(getUser());
-    }
-
-    @Override
-    protected void checkIfGameOver() {
-        if (userIsDestroyed()) {
-            loseGame();
-        } else if (boss.isDestroyed()) {
-            winGame();
-        }
-    }
-
-    @Override
-    protected void spawnEnemyUnits() {
-        if (getCurrentNumberOfEnemies() == 0) {
-            addEnemyUnit(boss);
-        }
-    }
-
-    @Override
-    protected LevelView instantiateLevelView() {
-        levelView = new LevelViewBoss(getRoot(), PLAYER_INITIAL_HEALTH);
-        return levelView;
-    }
-
-}
