--- src/main/java/com/example/demo/Levels/LevelParent.java
@@ -1,276 +1,0 @@
-package com.example.demo.Levels;
-
-import com.example.demo.Actors.ActiveActorDestructible;
-import com.example.demo.Planes.FighterPlane;
-import com.example.demo.LevelViews.LevelView;
-import com.example.demo.Planes.UserPlane;
-import javafx.animation.KeyFrame;
-import javafx.animation.Timeline;
-import javafx.scene.Group;
-import javafx.scene.Scene;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.input.KeyCode;
-import javafx.util.Duration;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-import java.util.Observable;
-
-public abstract class LevelParent extends Observable {
-
-    private static final double SCREEN_HEIGHT_ADJUSTMENT = 150;
-    private static final int MILLISECOND_DELAY = 50;
-    private final double screenHeight;
-    private final double screenWidth;
-    private final double enemyMaximumYPosition;
-
-    private final Group root;
-    private final Timeline timeline;
-    private final UserPlane user;
-    private final Scene scene;
-    private final ImageView background;
-
-    private final List<ActiveActorDestructible> friendlyUnits;
-    private final List<ActiveActorDestructible> enemyUnits;
-    private final List<ActiveActorDestructible> userProjectiles;
-    private final List<ActiveActorDestructible> enemyProjectiles;
-    private final LevelView levelView;
-    private int currentNumberOfEnemies;
-
-    public LevelParent(String backgroundImageName, double screenHeight, double screenWidth, int playerInitialHealth) {
-        this.root = new Group();
-        this.scene = new Scene(root, screenWidth, screenHeight);
-        this.timeline = new Timeline();
-        this.user = new UserPlane(playerInitialHealth);
-        this.friendlyUnits = new ArrayList<>();
-        this.enemyUnits = new ArrayList<>();
-        this.userProjectiles = new ArrayList<>();
-        this.enemyProjectiles = new ArrayList<>();
-
-        this.background = new ImageView(
-                new Image(Objects.requireNonNull(getClass().getResource(backgroundImageName)).toExternalForm()));
-        this.screenHeight = screenHeight;
-        this.screenWidth = screenWidth;
-        this.enemyMaximumYPosition = screenHeight - SCREEN_HEIGHT_ADJUSTMENT;
-        this.levelView = instantiateLevelView();
-        this.currentNumberOfEnemies = 0;
-        initializeTimeline();
-        friendlyUnits.add(user);
-    }
-
-    protected abstract void initializeFriendlyUnits();
-
-    protected abstract void checkIfGameOver();
-
-    protected abstract void spawnEnemyUnits();
-
-    protected abstract LevelView instantiateLevelView();
-
-    public Scene initializeScene() {
-        initializeBackground();
-        initializeFriendlyUnits();
-        levelView.showHeartDisplay();
-        return scene;
-    }
-
-    public void startGame() {
-        background.requestFocus();
-        timeline.play();
-    }
-
-    public void goToNextLevel(String levelName) {
-        timeline.stop();
-        timeline.getKeyFrames().clear();
-
-        background.setOnKeyPressed(null);
-        background.setOnKeyReleased(null);
-
-        root.getChildren().clear();
-
-        friendlyUnits.clear();
-        enemyUnits.clear();
-        userProjectiles.clear();
-        enemyProjectiles.clear();
-        setChanged();
-        notifyObservers(levelName);
-    }
-
-    private void updateScene() {
-        spawnEnemyUnits();
-        updateActors();
-        generateEnemyFire();
-        updateNumberOfEnemies();
-        handleEnemyPenetration();
-        handleUserProjectileCollisions();
-        handleEnemyProjectileCollisions();
-        handlePlaneCollisions();
-        removeAllDestroyedActors();
-        updateKillCount();
-        updateLevelView();
-        checkIfGameOver();
-    }
-
-    private void initializeTimeline() {
-        timeline.setCycleCount(Timeline.INDEFINITE);
-        KeyFrame gameLoop = new KeyFrame(Duration.millis(MILLISECOND_DELAY), e -> updateScene());
-        timeline.getKeyFrames().add(gameLoop);
-    }
-
-    private void initializeBackground() {
-        background.setFocusTraversable(true);
-        background.setFitHeight(screenHeight);
-        background.setFitWidth(screenWidth);
-        background.setOnKeyPressed(e -> {
-            KeyCode kc = e.getCode();
-            if (kc == KeyCode.UP)
-                user.moveUp();
-            if (kc == KeyCode.DOWN)
-                user.moveDown();
-            if (kc == KeyCode.LEFT)
-                user.moveLeft();
-            if (kc == KeyCode.RIGHT)
-                user.moveRight();
-            if (kc == KeyCode.SPACE)
-                fireProjectile();
-        });
-        background.setOnKeyReleased(e -> {
-            KeyCode kc = e.getCode();
-            if (kc == KeyCode.UP || kc == KeyCode.DOWN)
-                user.stopVerticalMovement();
-            if (kc == KeyCode.LEFT || kc == KeyCode.RIGHT)
-                user.stopHorizontalMovement();
-        });
-        root.getChildren().add(background);
-    }
-
-    private void fireProjectile() {
-        ActiveActorDestructible projectile = user.fireProjectile();
-        root.getChildren().add(projectile);
-        userProjectiles.add(projectile);
-    }
-
-    private void generateEnemyFire() {
-        enemyUnits.forEach(enemy -> spawnEnemyProjectile(((FighterPlane) enemy).fireProjectile()));
-    }
-
-    private void spawnEnemyProjectile(ActiveActorDestructible projectile) {
-        if (projectile != null) {
-            root.getChildren().add(projectile);
-            enemyProjectiles.add(projectile);
-        }
-    }
-
-    private void updateActors() {
-        friendlyUnits.forEach(ActiveActorDestructible::updateActor);
-        enemyUnits.forEach(ActiveActorDestructible::updateActor);
-        userProjectiles.forEach(ActiveActorDestructible::updateActor);
-        enemyProjectiles.forEach(ActiveActorDestructible::updateActor);
-    }
-
-    private void removeAllDestroyedActors() {
-        removeDestroyedActors(friendlyUnits);
-        removeDestroyedActors(enemyUnits);
-        removeDestroyedActors(userProjectiles);
-        removeDestroyedActors(enemyProjectiles);
-    }
-
-    private void removeDestroyedActors(List<ActiveActorDestructible> actors) {
-        List<ActiveActorDestructible> destroyedActors = actors.stream().filter(ActiveActorDestructible::isDestroyed)
-                .toList();
-        root.getChildren().removeAll(destroyedActors);
-        actors.removeAll(destroyedActors);
-    }
-
-    private void handlePlaneCollisions() {
-        handleCollisions(friendlyUnits, enemyUnits);
-    }
-
-    private void handleUserProjectileCollisions() {
-        handleCollisions(userProjectiles, enemyUnits);
-    }
-
-    private void handleEnemyProjectileCollisions() {
-        handleCollisions(enemyProjectiles, friendlyUnits);
-    }
-
-    private void handleCollisions(List<ActiveActorDestructible> actors1,
-                                  List<ActiveActorDestructible> actors2) {
-        for (ActiveActorDestructible actor : actors2) {
-            for (ActiveActorDestructible otherActor : actors1) {
-                if (actor.getBoundsInParent().intersects(otherActor.getBoundsInParent())) {
-                    actor.takeDamage();
-                    otherActor.takeDamage();
-                }
-            }
-        }
-    }
-
-    private void handleEnemyPenetration() {
-        for (ActiveActorDestructible enemy : enemyUnits) {
-            if (enemyHasPenetratedDefenses(enemy)) {
-                user.takeDamage();
-                enemy.destroy();
-            }
-        }
-    }
-
-    private void updateLevelView() {
-        levelView.removeHearts(user.getHealth());
-    }
-
-    private void updateKillCount() {
-        for (int i = 0; i < currentNumberOfEnemies - enemyUnits.size(); i++) {
-            user.incrementKillCount();
-        }
-    }
-
-    private boolean enemyHasPenetratedDefenses(ActiveActorDestructible enemy) {
-        return Math.abs(enemy.getTranslateX()) > screenWidth;
-    }
-
-    protected void winGame() {
-        timeline.stop();
-        levelView.showWinImage();
-    }
-
-    protected void loseGame() {
-        timeline.stop();
-        levelView.showGameOverImage();
-    }
-
-    protected UserPlane getUser() {
-        return user;
-    }
-
-    protected Group getRoot() {
-        return root;
-    }
-
-    protected int getCurrentNumberOfEnemies() {
-        return enemyUnits.size();
-    }
-
-    protected void addEnemyUnit(ActiveActorDestructible enemy) {
-        enemyUnits.add(enemy);
-        root.getChildren().add(enemy);
-    }
-
-    protected double getEnemyMaximumYPosition() {
-        return enemyMaximumYPosition;
-    }
-
-    protected double getScreenWidth() {
-        return screenWidth;
-    }
-
-    protected boolean userIsDestroyed() {
-        return user.isDestroyed();
-    }
-
-    private void updateNumberOfEnemies() {
-        currentNumberOfEnemies = enemyUnits.size();
-    }
-
-}
