--- src/main/java/com/example/demo/Planes/FighterPlane.java
@@ -1,40 +1,0 @@
-package com.example.demo.Planes;
-
-import com.example.demo.Actors.ActiveActorDestructible;
-
-public abstract class FighterPlane extends ActiveActorDestructible {
-
-    private int health;
-
-    public FighterPlane(String imageName, int imageHeight, double initialXPos, double initialYPos, int health) {
-        super(imageName, imageHeight, initialXPos, initialYPos);
-        this.health = health;
-    }
-
-    public abstract ActiveActorDestructible fireProjectile();
-
-    @Override
-    public void takeDamage() {
-        health--;
-        if (healthAtZero()) {
-            this.destroy();
-        }
-    }
-
-    protected double getProjectileXPosition(double xPositionOffset) {
-        return getLayoutX() + getTranslateX() + xPositionOffset;
-    }
-
-    protected double getProjectileYPosition(double yPositionOffset) {
-        return getLayoutY() + getTranslateY() + yPositionOffset;
-    }
-
-    private boolean healthAtZero() {
-        return health == 0;
-    }
-
-    public int getHealth() {
-        return health;
-    }
-
-}
