--- src/main/java/com/example/demo/Projectiles/BossProjectile.java
@@ -1,24 +1,0 @@
-package com.example.demo.Projectiles;
-
-public class BossProjectile extends Projectile {
-
-    private static final String IMAGE_NAME = "fireball.png";
-    private static final int IMAGE_HEIGHT = 75;
-    private static final int HORIZONTAL_VELOCITY = -15;
-    private static final int INITIAL_X_POSITION = 950;
-
-    public BossProjectile(double initialYPos) {
-        super(IMAGE_NAME, IMAGE_HEIGHT, INITIAL_X_POSITION, initialYPos);
-    }
-
-    @Override
-    public void updatePosition() {
-        moveHorizontally(HORIZONTAL_VELOCITY);
-    }
-
-    @Override
-    public void updateActor() {
-        updatePosition();
-    }
-
-}
