--- src/main/java/com/example/demo/Projectiles/EnemyProjectile.java
@@ -1,23 +1,0 @@
-package com.example.demo.Projectiles;
-
-public class EnemyProjectile extends Projectile {
-
-    private static final String IMAGE_NAME = "enemyFire.png";
-    private static final int IMAGE_HEIGHT = 32;
-    private static final int HORIZONTAL_VELOCITY = -10;
-
-    public EnemyProjectile(double initialXPos, double initialYPos) {
-        super(IMAGE_NAME, IMAGE_HEIGHT, initialXPos, initialYPos);
-    }
-
-    @Override
-    public void updatePosition() {
-        moveHorizontally(HORIZONTAL_VELOCITY);
-    }
-
-    @Override
-    public void updateActor() {
-        updatePosition();
-    }
-
-}
