--- src/main/java/com/example/demo/Projectiles/Projectile.java
@@ -1,19 +1,0 @@
-package com.example.demo.Projectiles;
-
-import com.example.demo.Actors.ActiveActorDestructible;
-
-public abstract class Projectile extends ActiveActorDestructible {
-
-    public Projectile(String imageName, int imageHeight, double initialXPos, double initialYPos) {
-        super(imageName, imageHeight, initialXPos, initialYPos);
-    }
-
-    @Override
-    public void takeDamage() {
-        this.destroy();
-    }
-
-    @Override
-    public abstract void updatePosition();
-
-}
