+++ src/main/java/com/example/demo/actors/ActiveActorDestructible.java
@@ -1,0 +1,33 @@
+package com.example.demo.actors;
+
+public abstract class ActiveActorDestructible extends ActiveActor implements Destructible {
+
+    private boolean isDestroyed;
+
+    public ActiveActorDestructible(String imageName, int imageHeight, double initialXPos, double initialYPos) {
+        super(imageName, imageHeight, initialXPos, initialYPos);
+        isDestroyed = false;
+    }
+
+    @Override
+    public abstract void updatePosition();
+
+    public abstract void updateActor();
+
+    @Override
+    public abstract void takeDamage();
+
+    @Override
+    public void destroy() {
+        setDestroyed();
+    }
+
+    protected void setDestroyed() {
+        this.isDestroyed = true;
+    }
+
+    public boolean isDestroyed() {
+        return isDestroyed;
+    }
+
+}
