+++ src/main/java/com/example/demo/actors/Destructible.java
@@ -1,0 +1,9 @@
+package com.example.demo.actors;
+
+public interface Destructible {
+
+    void takeDamage();
+
+    void destroy();
+
+}
