+++ src/main/java/com/example/demo/controller/Controller.java
@@ -1,15 +1,10 @@
 package com.example.demo.controller;
 
-import com.example.demo.Levels.LevelParent;
+import com.example.demo.levels.LevelParent;
 import javafx.scene.Scene;
-import javafx.scene.control.Alert;
-import javafx.scene.control.Alert.AlertType;
 import javafx.stage.Stage;
 
 import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
-import java.util.Observable;
-import java.util.Observer;
 
 public class Controller {
