+++ src/main/java/com/example/demo/images/ShieldImage.java
@@ -1,0 +1,35 @@
+package com.example.demo.images;
+
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+
+import java.util.Objects;
+
+public class ShieldImage extends ImageView {
+
+    private static final String IMAGE_NAME = "/com/example/demo/images/shield.png";
+    private static final int SHIELD_SIZE = 200;
+
+    public ShieldImage(double xPosition, double yPosition) {
+        this.setLayoutX(xPosition);
+        this.setLayoutY(yPosition);
+        this.setImage(new Image(Objects.requireNonNull(getClass().getResource(IMAGE_NAME)).toExternalForm()));
+        this.setVisible(false);
+        this.setFitHeight(SHIELD_SIZE);
+        this.setFitWidth(SHIELD_SIZE);
+    }
+
+    public void showShield() {
+        this.setVisible(true);
+    }
+
+    public void hideShield() {
+        this.setVisible(false);
+    }
+
+    public void updatePosition(double x, double y) {
+        setLayoutX(x);
+        setLayoutY(y);
+    }
+
+}
