+++ src/main/java/com/example/demo/projectiles/Projectile.java
@@ -1,0 +1,19 @@
+package com.example.demo.projectiles;
+
+import com.example.demo.actors.ActiveActorDestructible;
+
+public abstract class Projectile extends ActiveActorDestructible {
+
+    public Projectile(String imageName, int imageHeight, double initialXPos, double initialYPos) {
+        super(imageName, imageHeight, initialXPos, initialYPos);
+    }
+
+    @Override
+    public void takeDamage() {
+        this.destroy();
+    }
+
+    @Override
+    public abstract void updatePosition();
+
+}
