+++ src/main/java/com/example/demo/projectiles/UserProjectile.java
@@ -1,0 +1,23 @@
+package com.example.demo.projectiles;
+
+public class UserProjectile extends Projectile {
+
+    private static final String IMAGE_NAME = "userfire.png";
+    private static final int IMAGE_HEIGHT = 6;
+    private static final int HORIZONTAL_VELOCITY = 15;
+
+    public UserProjectile(double initialXPos, double initialYPos) {
+        super(IMAGE_NAME, IMAGE_HEIGHT, initialXPos, initialYPos);
+    }
+
+    @Override
+    public void updatePosition() {
+        moveHorizontally(HORIZONTAL_VELOCITY);
+    }
+
+    @Override
+    public void updateActor() {
+        updatePosition();
+    }
+
+}
